Disable DHCP support and fixup MAC address discovery
Getting the value of option 245 out of the dhcpcd lease file will be a
pain. Since that's optional, I've just set DHCP to be unavailable on
NetBSD for the moment.

There were some Linuxisms in the MAC address discovery code. We
probably don't even need that code while DHCP is set to be
unavailable.